+++ Server/Server/Models/Messaging/Handshake.cs
@@ -7,7 +7,7 @@
     {
         public Handshake() : base()
         {
-            MessageType = MessageTypes.Handshake;
+            MessageType = (int) MessageTypes.Handshake;
         }
     }
 }
